+++ src/misc/lru_map.rs
@@ -74,6 +74,7 @@ where
     }
 }
 
+#[cfg(test)]
 mod tests {
 
     #[test]